pyproject.toml
[project]
name = "r1-arc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
# Core dependencies that are needed everywhere
dependencies = [
    "anyio>=4.8.0",
    "asyncio>=3.4.3",
    "better-exceptions>=0.3.3",
    "deepseek-tokenizer>=0.1.3",
    "devtools>=0.12.2",
    "hf-transfer>=0.1.9",
    "ipykernel>=6.29.5",
    "jinja2>=3.1.6",
    "jupyter>=1.1.1",
    "libcst>=1.6.0",
    "matplotlib>=3.10.1",
    "openai>=1.66.1",
    "orjson>=3.10.15",
    "pygments>=2.19.1",
    "pytest-xdist>=3.6.1",
    "python-dotenv>=1.0.1",
    "scipy>=1.15.2",
    "setuptools>=76.0.0",
    "tenacity>=9.0.0",
    "tiktoken>=0.9.0",
    "typer>=0.15.2",
    "viztracer>=1.0.2",
]

[project.packages.find]
where = ["src"]

[dependency-groups]
# Development dependencies
dev = [
    "pytest>=8.3.5",
    "syrupy>=4.9.0",
    "pytest-asyncio>=0.25.3",
    "datasets>=3.3.2",
]

# Remote CUDA Linux dependencies
remote = [
    "gradio>=5.20.1",
    "torch>=2.5.1",
    "transformers>=4.49.0",
    "accelerate==1.4.0",
    "bitsandbytes==0.45.3",
    "gradio>=5.20.1",
    "wandb>=0.19.8",
    "msgspec>=0.19.0",
    "peft==0.14.0",
    "torch>=2.5.1",
    "transformers>=4.49.0",
    "trl==0.15.2",
    "unsloth==2025.3.10",
    "unsloth-zoo==2025.3.9",
    "vllm==0.7.3",
]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["src"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-p no:warnings"
asyncio_default_fixture_loop_scope = "function"
